Improve error handling
When given invalid syntax, it can be difficult to find out where the
problem is given only the position without the context of the input
string.

We define a new, specific EbnfException class storing the position of
the parsing error and throw that instead of a regular Exception.

This allows the exception handler to provide some context for the error
as a short substring of the input text.